Fixed minor missed stdint and stddef

+++ src/arch/x86/time/time.c
@@ -1,8 +1,8 @@
 #include "time.h"
 #include "rtc.h"
+#include "types.h"
 
 #include <stdbool.h>
-#include <stdint.h>
 
 time_t volatile seconds_since_epoch = 0;
 
@@ -15,24 +15,24 @@ __attribute__((warn_unused_result)) inline time_t getepoch(void)
 
 inline void setepoch(time_t new) { seconds_since_epoch = new; }
 
-static inline int32_t is_leap(uint32_t year)
+static inline s32 is_leap(u32 year)
 {
     return (year % 4 == 0 && year % 100 != 0) || (year % 400 == 0);
 }
 
 time_t to_epoch(rtc_time_t* t)
 {
-    uint64_t epoch_seconds;
-    uint32_t full_year = 2000 + t->year;
-    uint32_t days_since_epoch = 0;
-    uint32_t const days_in_month[] = { 0, 31, 28, 31, 30, 31, 30,
+    unsigned long long epoch_seconds;
+    u32 full_year = 2000 + t->year;
+    u32 days_since_epoch = 0;
+    u32 const days_in_month[] = { 0, 31, 28, 31, 30, 31, 30,
         31, 31, 30, 31, 30, 31 };
 
-    for (uint32_t y = 1970; y < full_year; ++y) {
+    for (u32 y = 1970; y < full_year; ++y) {
         days_since_epoch += is_leap(y) ? 366 : 365;
     }
 
-    for (uint8_t m = 1; m < t->month; ++m) {
+    for (u8 m = 1; m < t->month; ++m) {
         days_since_epoch += days_in_month[m];
         if (m == 2 && is_leap(full_year)) {
             days_since_epoch++;
@@ -48,17 +48,17 @@ time_t to_epoch(rtc_time_t* t)
 
 void from_epoch(time_t epoch, rtc_time_t* t)
 {
-    uint32_t const days_in_month[] = { 0, 31, 28, 31, 30, 31, 30,
+    u32 const days_in_month[] = { 0, 31, 28, 31, 30, 31, 30,
         31, 31, 30, 31, 30, 31 };
-    uint32_t current_year = 1970;
-    uint32_t days_in_current_year;
+    u32 current_year = 1970;
+    u32 days_in_current_year;
 
-    uint64_t seconds_of_day = epoch % 86400ULL;
+    unsigned long long seconds_of_day = epoch % 86400ULL;
     t->hour = seconds_of_day / 3600;
     t->minute = (seconds_of_day % 3600) / 60;
     t->second = seconds_of_day % 60;
 
-    uint64_t total_days = epoch / 86400ULL;
+    unsigned long long total_days = epoch / 86400ULL;
 
     while (true) {
         days_in_current_year = is_leap(current_year) ? 366 : 365;
